requirements.txt
streamlit==1.29.0
PyPDF2
openai
python-docx